System clarity tweaks
