+++ Content/Forest/Cartography/Cartographer.cs
@@ -13,7 +13,6 @@ using SpiritReforged.Content.Forest.Cartography.Pins;
 using SpiritReforged.Content.Forest.Cartography.Maps;
 using SpiritReforged.Common.ItemCommon;
 using SpiritReforged.Common.EmoteCommon;
-using Terraria.GameContent.UI;
 
 namespace SpiritReforged.Content.Forest.Cartography;
 
@@ -31,7 +30,7 @@ public class Cartographer : WorldNPC, ITravelNPC
 		return cartographer;
 	}
 
-	public override void Load() => CustomEmote.LoadCustomEmote(new(GetType(), EmoteID.Category.Town, () => NPC.AnyNPCs(ModContent.NPCType<Cartographer>())));
+	public override void Load() => AutoEmote.LoadFaceEmote(this, () => NPC.AnyNPCs(ModContent.NPCType<Cartographer>()));
 	public override void SetStaticDefaults()
 	{
 		base.SetStaticDefaults();
